Initial implementation of the client side code.
Implements initial code for a user interface for comments in the re-frame framework
for clojurescript. Based on the leiningen template for re-frame apps it divides the
code into subscriptions, views, events, database, core (bootstrapping), and config
modules. This should make it relatively easy to grok.

The piece of code re-frame app currently supports fetching and displaying comment
threads. Comment threads are stored and displayed as a recursive structure.

Currently only supports posting top level comments. Replying to existing comments is
on the todo-list, but should be doable. It only displays comment content as pure text.
The goal is to support a limited part of markdown syntax to allow rich comments.

Uses external avatar service http://avatars.adorable.io/ to provide recognizable avatars.
While I initially did not feel using external avatar services was a good fit for a privacy
minded comment system, this service appears to behave nicely in the privacy regard.
It is also open source which means that I (or a user) could host it themselves if they
so want.

Signed-off-by: Snorre Magnus Davøen <snorremd@gmail.com>